feat(loads-context): improved code splitting into files
