+++ src/utils/loads-context/calculate-total-pages/index.ts
@@ -1,0 +1,3 @@
+export const calculateTotalPages = (totalItems: number, itemsPerPage: number) => {
+  return Math.ceil(totalItems / itemsPerPage)
+}
